requirements.txt
envs==1.4
filelock==3.8.0
flatbuffers==20181003210633
gast==0.4.0
gdown==4.5.1
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.6.0
idna==3.4
importlib-metadata==5.0.0
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mock==4.0.3
numpy==1.23.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pre-commit==3.0.2
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smbus==1.1.post2
soupsieve==2.3.2.post1
sqlalchemy==1.4.41
tb-mqtt-client==1.4.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==2.0.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.9.0
